# external/interface/httpintf/swagger/api.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to provide API over HTTP to toggler service,\nin which you can interact with the service in a programmatic way.",
    "title": "provides API on HTTP layer to the toggler service.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "version": "0.2.0"
  },
  "basePath": "/api",
  "paths": {
    "/deployment-environments": {
      "get": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployment",
          "environment"
        ],
        "summary": "List all the deployment environment that can be used to manage a feature rollout.",
        "operationId": "listDeploymentEnvironments",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listDeploymentEnvironmentResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "description": "This operation allows you to create a new deployment environment.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployment",
          "environment"
        ],
        "summary": "Create a deployment environment that can be used for managing a feature rollout.",
        "operationId": "createDeploymentEnvironment",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "environment": {
                  "$ref": "#/definitions/Environment"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createDeploymentEnvironmentResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/deployment-environments/{envID}": {
      "put": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployment",
          "environment"
        ],
        "summary": "Update a deployment environment.",
        "operationId": "updateDeploymentEnvironment",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "EnvironmentID",
            "description": "EnvironmentID is the deployment environment id or the alias name.",
            "name": "envID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "environment": {
                  "$ref": "#/definitions/Environment"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateDeploymentEnvironmentResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
    "/release-flags": {
      "get": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "List all the release flag that can be used to manage a feature rollout.",
        "operationId": "listReleaseFlags",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listReleaseFlagResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "description": "This operation allows you to create a new release flag.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "Create a release flag that can be used for managing a feature rollout.",
        "operationId": "createReleaseFlag",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "flag": {
                  "$ref": "#/definitions/Flag"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createReleaseFlagResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Environment": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/domains/deployment"
    },
    "Error": {
      "description": "Error contains the details of the error",
      "type": "object",
      "properties": {
        "code": {
          "description": "The constant code of the error that can be used for localisation",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 401
        },
        "message": {
          "description": "The message that describe the error to the developer who do the integration.\nNot meant to be propagated to the end-user.\nThe Message may change in the future, it it helps readability,\nplease do not rely on the content in any way other than just reading it.",
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/external/interface/httpintf/httpapi"
    },
    "Flag": {
      "description": "Flag is the basic entity with properties that feature flag holds",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/domains/release"
    },
...
  },
  "responses": {
    "createDeploymentEnvironmentResponse": {
      "description": "CreateDeploymentEnvironmentResponse",
      "schema": {
        "type": "object",
        "properties": {
          "environment": {
            "$ref": "#/definitions/Environment"
          }
        }
      }
    },
    "createReleaseFlagResponse": {
      "description": "CreateReleaseFlagResponse",
      "schema": {
        "type": "object",
        "properties": {
          "flag": {
            "$ref": "#/definitions/Flag"
          }
        }
      }
    },
...
    "errorResponse": {
      "description": "ErrorResponse will contains a response about request that had some kind of problem.\nThe details will be included in the body.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
...
    "listDeploymentEnvironmentResponse": {
      "description": "ListDeploymentEnvironmentResponse",
      "schema": {
        "type": "object",
        "properties": {
          "environments": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Environment"
            },
            "x-go-name": "Environments"
          }
        }
      }
    },
    "listReleaseFlagResponse": {
      "description": "ListReleaseFlagResponse",
      "schema": {
        "type": "object",
        "properties": {
          "flags": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Flag"
            },
            "x-go-name": "Flags"
          }
        }
      }
    },
...
    "updateDeploymentEnvironmentResponse": {
      "description": "UpdateDeploymentEnvironmentResponse",
      "schema": {
        "type": "object",
        "properties": {
          "environment": {
            "$ref": "#/definitions/Environment"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "AppKey": {
      "type": "apiKey",
      "name": "X-APP-KEY",
      "in": "header"
    },
    "AppToken": {
      "type": "apiKey",
      "name": "X-APP-TOKEN",
      "in": "header"
    }
  },
  "security": [
    {
      "AppKey": [
        "[]"
      ]
    }
  ]
}